common/SubProcess: Move definitions to cc file, plus trivial cleanup
